% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_countries.R
\name{get_all_countries}
\alias{get_all_countries}
\title{Fetch All Countries}
\usage{
get_all_countries()
}
\value{
A data frame containing the list of all countries.
}
\description{
Retrieve all countries for which DTM data is publicly available through the API.
}
\examples{
\dontshow{if (!identical(Sys.getenv("DTM_SUBSCRIPTION_KEY"), "")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
countries_df <- get_all_countries()
head(countries_df)
\dontshow{\}) # examplesIf}
}
