% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_operations.R
\name{get_all_operations}
\alias{get_all_operations}
\title{Fetch All Operations}
\usage{
get_all_operations()
}
\value{
A data frame containing the list of all operations.
}
\description{
Retrieve all operations for which DTM data is publicly available through the API.
}
\examples{
\dontrun{
# Fetch all operations
operations_df <- get_all_operations()
head(operations_df)
}
}
